Require illuminate/support instead of the full laravel framework.
requiring laravel/framework was causing this package to pull in the entire laravel framework instead of just the libraries needed to work. This was causing errors when pulling this package into OctoberCMS which is based on laravel.

Set minimum version contstraints for dev packages so we don't pull in the oldest version of orchestra/testbench when running composer install.